Create repeatWhile.js
Write a function named repeat that takes two arguments: a string and a number. It should return a string that is the string that you sent into the function, the number of times that was specified. So if you called the function like: repeat("x", 5) the return value should be: "xxxxx". If you did: repeat("foo", 3) the return should be: "foofoofoo". Do this without using the built-in string.repeat() method. (Which is good to know about, but don't use it here.)​